Additional error checking.

+++ main.go
@@ -3,27 +3,60 @@ package main
 import (
 	"fmt"
 	"io/ioutil"
-	"log"
 	"os"
 	"path"
+
+	"flag"
 
 	"github.com/rainycape/magick"
 )
 
-func main() {
-	var dir = "/home/am/projects/sdk/images/gallery"
+var source string
+var dest string
+var width int
 
-	resizeFilesInDir(dir)
+func init() {
+
+	flag.StringVar(&source, "src", "", "Directory containing files that we want to convert")
+	flag.StringVar(&dest, "dst", "", "Directory we want to output to")
+	flag.IntVar(&width, "width", 200, "Width to resize to")
 }
 
-func resizeFilesInDir(dir string) {
+func main() {
+	//var dir = "/home/am/projects/sdk/images/gallery"
+	fmt.Println("rar:", flag.Args())
+	flag.Parse()
+
+	fmt.Println("rar: ", source)
+	fmt.Println("foo: ", dest)
+	fmt.Println("tail:", flag.Args())
+
+	// args := flag.Args()
+	// fmt.Printf("rar %d", len(args))
+
+	/*
+		if len(args) < 1 {
+
+			flag.Usage()
+			os.Exit(2)
+		}
+	*/
+	resizeFilesInDir(source, dest, width)
+}
+
+func resizeFilesInDir(dir string, dest string, size int) {
 	filesInDir, err := ioutil.ReadDir(dir)
 	if err != nil {
 		fmt.Print(err)
 	}
 	if len(filesInDir) < 1 {
 		fmt.Print("Directory contains no files")
-		os.Exit(0)
+		os.Exit(2)
+	}
+
+	if size < 1 {
+		fmt.Printf("Size : {size} must be > 0")
+		os.Exit(2)
 	}
 
 	for _, fileName := range filesInDir {
@@ -59,8 +92,8 @@ func resizeFile(fileName string) {
 		newHeight = -1
 		layout = "portrait"
 	} else {
-		newWidth = -1
-		newHeight = 200
+		newWidth = 200
+		newHeight = 1
 		layout = "landscape"
 	}
 
@@ -74,19 +107,52 @@ func resizeFile(fileName string) {
 	outputFilename := path.Base(fileName)
 	newFilename := path.Join("/tmp/images", layout, outputFilename)
 
-	writeImage(newFilename, newImage)
+	var imgWritten bool
+
+	imgWritten, err = writeImage(newFilename, newImage)
+
+	if !imgWritten {
+		if err != nil {
+			fmt.Print(err)
+		}
+	}
 
 	fmt.Printf("File: %s has been resized to %d x %d \n", newFilename, newImage.Width(), newImage.Height())
 }
 
-func writeImage(newFilename string, newImage *magick.Image) {
-	fo, err := os.Create(newFilename)
+func writeImage(newFilename string, newImage *magick.Image) (bool, error) {
+
+	doesExist, err := exists(path.Base(newFilename))
+
+	// @todo use case
 	if err != nil {
-		log.Fatal(err)
+		return false, err
+	}
+
+	if !doesExist {
+		return false, nil
+	}
+
+	fo, err := os.Create(newFilename)
+
+	if err != nil {
+		return false, err
 	}
 
 	defer fo.Close()
 	newImage.Encode(fo, magick.NewInfo())
+	return true, nil
+}
+
+func exists(path string) (bool, error) {
+	_, err := os.Stat(path)
+	if err == nil {
+		return true, nil
+	}
+	if os.IsNotExist(err) {
+		return false, nil
+	}
+	return true, err
 }
 
 // ShowFormats displays all the formats supported
